+++ Principal/.gitignore
@@ -1,0 +1,8 @@
+/Projectile.class
+/Background.class
+/Enemy.class
+/Player.class
+/Main.class
+/GameLib.class
+/MyFrame.class
+/MyKeyAdapter.class
