+++ Principal/GameLib.java
@@ -1,3 +1,4 @@
+package Principal;
 import java.awt.Color;
 import java.awt.Graphics;
 import java.awt.Toolkit;
