+++ Principal/Main.java
@@ -1,3 +1,4 @@
+package Principal;
 import java.awt.Color;
 
 /***********************************************************************/
